+++ chemicals/phase_change.py
@@ -252,7 +252,7 @@ else:
 
 Tb_all_methods = (miscdata.HEOS, CRC_INORG, CRC_ORG, miscdata.COMMON_CHEMISTRY,
                   miscdata.WEBBOOK, YAWS, miscdata.WIKIDATA, miscdata.JOBACK)
-"""Tuple of method name keys. See the `Tb` for the actual references"""
+"""Tuple of method name keys. See the `Tbg` for the actual references"""
 
 @mark_numba_incompatible
 def Tb_methods(CASRN):
@@ -315,6 +315,9 @@ def Tb(CASRN, method=None):
         * 'YAWS', a large compillation of data from a
           variety of sources both experimental and predicted;
           no data points are sourced in the work of [2]_.
+        * 'HEOS', a series of values from the NIST REFPROP Database for
+          Highly Accurate Properties of Industrially Important Fluids
+          (and other high-precision fundamental equations of state)
 
     Examples
     --------
@@ -340,6 +343,10 @@ def Tb(CASRN, method=None):
        doi:10.1080/00986448708960487.
     .. [6] Shen, V.K., Siderius, D.W., Krekelberg, W.P., and Hatch, H.W., Eds.,
        NIST WebBook, NIST, http://doi.org/10.18434/T4M88Q
+    .. [7] Huber, Marcia L., Eric W. Lemmon, Ian H. Bell, and Mark O. McLinden. 
+       "The NIST REFPROP Database for Highly Accurate Properties of Industrially
+       Important Fluids." Industrial & Engineering Chemistry Research 61, no. 42 
+       (October 26, 2022): 15449-72. https://doi.org/10.1021/acs.iecr.2c01427.
     '''
     if dr.USE_CONSTANTS_DATABASE and method is None:
         val, found = database_constant_lookup(CASRN, 'Tb')